@@ -5,19 +5,19 @@
   "description": "A simple starter to get up and developing quickly with Gatsby",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
-    "@telekom/scale-components-react-neutral": "^3.0.0-beta.1",
-    "@telekom/scale-components-neutral": "^3.0.0-beta.1",
-    "gatsby": "^2.20.36",
-    "gatsby-image": "^2.3.5",
-    "gatsby-plugin-manifest": "^2.3.7",
-    "gatsby-plugin-offline": "^3.1.5",
-    "gatsby-plugin-react-helmet": "^3.2.5",
-    "gatsby-plugin-sharp": "^2.5.7",
-    "gatsby-source-filesystem": "^2.2.5",
-    "gatsby-transformer-sharp": "^2.4.7",
+    "@telekom/scale-components-neutral": "^3.0.0-beta.117",
+    "@telekom/scale-components-react-neutral": "^3.0.0-beta.117",
+    "gatsby": "^5.0.1",
+    "gatsby-plugin-image": "^3.0.0",
+    "gatsby-plugin-manifest": "^5.0.0",
+    "gatsby-plugin-offline": "^6.0.0",
+    "gatsby-plugin-react-helmet": "^6.0.0",
+    "gatsby-plugin-sharp": "^5.0.0",
+    "gatsby-source-filesystem": "^5.0.0",
+    "gatsby-transformer-sharp": "^5.0.0",
     "prop-types": "^15.7.2",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.0.0",
     "react-table": "^7.0.4"
   },
