@@ -3,13 +3,13 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@telekom/scale-components-neutral": "^3.0.0-beta.16",
-    "@telekom/scale-components-react-neutral": "^3.0.0-beta.16",
+    "@telekom/scale-components-neutral": "^3.0.0-beta.117",
+    "@telekom/scale-components-react-neutral": "^3.0.0-beta.117",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.0.1"
   },
