@@ -4,14 +4,14 @@
   "private": true,
   "main": "dist/index.js",
   "dependencies": {
-    "@telekom/scale-components-neutral": "^3.0.0-beta.1",
-    "@telekom/scale-components-react-neutral": "^3.0.0-beta.1",
+    "@telekom/scale-components-neutral": "^3.0.0-beta.117",
+    "@telekom/scale-components-react-neutral": "^3.0.0-beta.117",
     "@types/jest": "25.1.4",
     "@types/node": "13.9.8",
-    "@types/react": "16.9.31",
-    "@types/react-dom": "16.9.6",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "3.4.1"
   },
   "devDependencies": {
