@@ -8,9 +8,9 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@telekom/scale-components-neutral": "^3.0.0-beta.16",
+    "@telekom/scale-components-neutral": "^3.0.0-beta.117",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11"
+    "vue": "^2.7.10"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
