@@ -1,7 +1,6 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import { defineCustomElements } from "@telekom/scale-components-neutral/loader";
-
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import "@telekom/scale-components-neutral/dist/scale-components/scale-components.css";
 
@@ -10,12 +9,9 @@ import reportWebVitals from './reportWebVitals';
 
 defineCustomElements(window);
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+const container = document.getElementById('root');
+const root = createRoot(container); // createRoot(container!) if you use TypeScript
+root.render(<App/>);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
