@@ -1,4 +1,4 @@
-const { defineCustomElements } = require("@telekom/scale-components/loader");
+const { defineCustomElements } = require("@telekom/scale-components-neutral/loader");
 
 exports.onInitialClientRender = ({ element }, options) => {
   defineCustomElements(window);
