@@ -1,8 +1,15 @@
 <template>
-  <div>
-    <scale-button>foobar</scale-button>
+  <div style="display:flex; flex-direction: column; width: 100%; height: 100vh; justify-content: center; align-items: center; ">
+      <h1>Scale + Nuxt 3</h1>
+      <scale-button
+        href="https://www.brand-design.telekom.com/en/account/login/?redirect_url=/scale/"
+        >Open the docs</scale-button
+      >
   </div>
 </template>
 <style>
-@import "@telekom/scale-components/dist/scale-components/scale-components.css";
+@import "@telekom/scale-components-neutral/dist/scale-components/scale-components.css";
+:root {
+  font-family: sans-serif;
+}
 </style>
