@@ -1,8 +1,7 @@
-import { defineNuxtConfig } from 'nuxt'
 // https://v3.nuxtjs.org/api/configuration/nuxt.config
 export default defineNuxtConfig({
   build: {
-    transpile: ['@telekom/scale-components']
+    transpile: ['@telekom/scale-components-neutral']
   },  
   vue: {
     compilerOptions: {
