@@ -1,5 +1,5 @@
 /* @vite-ignore */
-import  {defineCustomElements, applyPolyfills} from "@telekom/scale-components/loader"
+import  {defineCustomElements, applyPolyfills} from "@telekom/scale-components-neutral/loader"
 
 export default defineNuxtPlugin(nuxtApp => {
     if (process.client) {
